feedback fixes
